bugfixes - see note below
couple of major bugfixes here. firstly f(t) was not always being
correctly zeroed, which i think would tend to allow for generation of
excitations long after the pulse is over. second, there was a typo in
one of the calls to possible_move which meant the wrong site was being
selected, meaning that legitimate moves were not being selected