fix intra rate setup and annihilation behaviour
fortran's column major so when you're reshaping a 1D array to 2D it goes
in column first! i knew this and yet forgot to transpose it. this was
making internal transfer between states behave weirdly (since that's the
only asymmetric matrix we're dealing with atm) and there was also a bug
in the annihilation rate constant which was allowing for events that
shouldn't have been possible. fixed now, all seems to be working